+++ app/src/test/java/com/github/houseorganizer/houseorganizer/ExpenseUnitTest.java
@@ -15,7 +15,7 @@ import java.util.Objects;
 
 public class ExpenseUnitTest {
 
-    HashMap<String, Double> map = new HashMap<>();
+    HashMap<String, Float> map = new HashMap<>();
     Expense expense = new Expense("title", 0, "payee", map);
 
     @Test
